custom-elements/creating-and-passing-registries/share-registry-import-document.html

<!DOCTYPE html>
<html>
<head>
<title>When creating an import, use the registry of the master document</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="When creating an import, use the registry of the master document">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#creating-and-passing-registries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

testInIFrame('../resources/import-master.html', function(doc) {
    var link = doc.querySelector('link[rel=import]');
    var doc2 = link.import;
    var name = 'x-frame';
    doc.registerElement(name);
    assert_throws(
        'NotSupportedError',
        function () { doc2.registerElement(name); },
        'Registering a custom element type name that is already registered in a shared ' +
            'registry should throw an exception');
}, 'Registry of the import should share master document registry');


testInIFrame('../resources/import-master.html', function(doc) {
    var link = doc.querySelector('link[rel=import]');
    var doc2 = link.import;
    var name = 'x-frame';
    doc2.registerElement(name);
    assert_throws(
        'NotSupportedError',
        function () { doc.registerElement(name); },
        'Registering a custom element type name that is already registered in a shared ' +
            'registry should throw an exception');
}, 'Registry of the master document should be shared with imported document\'s registry');

</script>
</body>
</html>
